requirements.txt
openai==1.6.1
httpx==0.27.0
python-dotenv==1.0.1
datasets>=2.16.0
huggingface-hub>=0.20.0
matplotlib>=3.7.0